Move tests to sub-directory: poly-commitment
