--- poly-commitment/src/tests/mod.rs
@@ -1,3 +1,0 @@
-mod batch_15_wires;
-mod commitment;
-mod serialization;
