+++ poly-commitment/tests/lib.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+pub mod commitment;
